requirements.txt
agno==1.5.9
pandas>=2.0.0
numpy>=1.24.0
google-genai>=1.0.0
lancedb>=0.1.0
ipykernel
pathlib>=1.0.1
sentence-transformers
aiofiles
gradio>=4.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0